Adds handler which ignores modifiers submitted as keys
Some setups provide our environment with a keyDown for simple modifier
keys being pressed.  We assumed these were actions we did not
understand and therefore let the content-editable handle these
actions.  This makes little sense for modifiers, their effects are
submitted when a subsequent key is pressed.

We have chosen to pass the event to other handlers should they desired
to do so, as they maybe relevant to register for click handlers and
the likes.