Replace NodeWalker implementation with marawa's
The NodeWalker in this component can be replaced with marawa's.
There's a small difference in execution, being that the create method
does not exist in traditional classes.  Because we suspect plugins may
import the implementation from here, rather than from other places, we
offer a create interface to allow them to continue importing as they
do now.