Increasing timeout of generateDiffEvents to 320ms
The increased timout allows fast typers to type their text without the
editor blocking in between.  Slower typers will not notice the
freeze.

The timeout would ideally be calculated based on user input.  We could
better estimate when to trigger when we'd know how faster users are
typing.  Slow users will not notice the slow effect so they can
receive the input at all times, fast typers likely want to input to
appear once they stopped typing briefly.  This is a future story that
should be tackled in a reusable manner.