Optimize getRichNodeMatchingDomNode
The flatMap used in getRichNodeMatchingDomNode call could be further
optimized.  The Ember.get call that was in there is fairly costly due
to support for older Ember versions and the likes.  The new getters
are notably faster on native objects.  Furhter, the flatMap method has
support for fetching only the first matching element if that is
found.  This code-path is now followed in the current code, greatly
optimizing the speed of the consuming code.