Refactor: the *ListAction is same logic for both ul and li
Move them to a common procedure.
Note: the created handleListaction function signature is very dependent
on how the wiring is currently implemented. So this is likely to
change.
The logic inside however is considered (for now) as more stable.

Note 2: a similar refactor will happen for indent/unindent in a later commit.