Add experimental attributes for RDFa
Like we have the data-editor-position-level, we now also have the
data-editor-rdfa-position-level.  This implementation should be
sufficient to get some information from the editor towards the user
with respect to where they are.

The future evolution of this component is to call a before/after
helper on node changes, allowing for aspects to hook into this
behaviour.  That would allow us to move this code into the
ember-rdfa-editor instead which makes for a much better abstraction.