prefer function declarations over function expressions
with function declarations, functions can be ordered as we seem fit. with
function expressions (using the const keyword) functions are inaccessible until
execution reaches it.

on the flip side, function expressions using const are immutable whereas
function declarations are not. So using const might be useful if we want to make
absolutely sure functions can't be redefined.