improve cursor movement when a hint is applied
previously we only checked if the cursor was within the range of the highlight.
but this does not suffice. If we adjust a dom node that has the cursor in it we
need to reposition the cursor so the browser displays it correctly.

Now always repositioning the cursor to be on the safe side.

A more fine grained alternative would be to compare the range of the highlight
and the current node. if these overlap the cursor must be repositioned.